♻️ JE-83-refactor: 이미지 최적화
