+++ src/pages/_document.tsx
@@ -3,8 +3,9 @@ import { Head, Html, Main, NextScript } from 'next/document';
 export default function Document() {
   return (
     <Html lang="ko">
-      <Head />
-      <link rel="icon" href="/logo/tripzipFavicon.png" type="image/png" />
+      <Head>
+        <link rel="icon" href="/logo/tripzipFavicon.webp" type="image/webp" />
+      </Head>
       <body>
         <Main />
         <NextScript />
